fix: oopsie on not including dist/ folder

+++ packages/gleamy/scripts/postpack.cjs
@@ -6,6 +6,9 @@ const result = JSON.parse(fs.readFileSync(appRoot + '/entitiesToPack.json'));
 console.log('Gleamy: cleaning up.');
 try {
   result.entitiesToPack.forEach((entity) => {
+    if (entity === "dist") {
+      return void(0);
+    }
     fs.rmSync(`../../packages/gleamy/${entity}`, { recursive: true });
   });
 } catch (error) {
